refactor(migration): gather every database-level object diff in one module
diff_buckets and diff_analyzers move to migration::diff_objects, which the
module doc already described, and are re-exported from migration::diff so
no consumer path changes.

migration::diff now sits below the size it started this branch at, despite
gaining change-feed and view diffing.
